requirements.txt
streamlit
lark-parser
graphviz
networkx
matplotlib
pandas


